add logic of manage vouchers

+++ app/controllers/payments_controller.rb
@@ -6,7 +6,7 @@ class PaymentsController < ApplicationController
 
   def at_sale
     # Destroy payments of a payment method if seller passes no amount
-    if params[:payment_amount].empty?
+    if !params[:payment_voucher] && params[:payment_amount].empty?
       Sale.find(params[:payment_sale_id]).payments.where(:payment_method_id => params[:payment_method_id]).destroy_all
     else
       # Destroy previous payments of a payment method (if any)
@@ -14,14 +14,18 @@ class PaymentsController < ApplicationController
       if prev_payments.count > 0
         prev_payments.destroy_all
       end
-      @payment = Payment.new(
-        :sale_id => params[:payment_sale_id],
-        :amount => params[:payment_amount],
-        :payment_method_id => params[:payment_method_id]
-      )
-      @payment.save
+      if (params[:payment_voucher] && !params[:payment_voucher].blank? ) || params[:payment_amount]
+        @payment = Payment.new(
+          :sale_id => params[:payment_sale_id],
+          :amount => params[:payment_voucher].blank? ? params[:payment_amount] : Voucher.find(params[:payment_voucher]).amount,
+          :payment_method_id => params[:payment_method_id]
+        )
+        @payment.save
+        Voucher.find(params[:payment_voucher]).update_attributes(:payment_id => @payment.id, :state => "canjeado") if params[:payment_voucher]
+      end
     end
     @sale = Sale.find(params[:payment_sale_id])
+
     hobo_ajax_response if request.xhr?
   end
